Fixed compile warnings.
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@9476 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ idlib/Str.cpp
@@ -945,7 +945,7 @@ idStr idStr::Join( const idList<idStr> &tokens, const char *separator )
 		return "";
 
 	idStr res = tokens[0];
-	for (size_t i = 1; i < tokens.Num(); i++) {
+	for (int i = 1; i < tokens.Num(); i++) {
 		res += separator;
 		res += tokens[i];
 	}
